Refactor manifest module structure and clean up redundant comments
• Split manifest module into types.rs and loader.rs submodules
• Move filesystem tests from unit tests to integration tests
• Remove redundant comments that restate obvious information
• Improve Epoch documentation with distributed systems context
